Create ES6 modules for the project and put each class in seprate module
