+++ src/modules/ListProperties.js
@@ -1,0 +1,7 @@
+export default class ListProperties {
+  constructor(description, completed, index) {
+    this.description = description;
+    this.completed = completed;
+    this.index = index;
+  }
+}
